Consolidate i18n config in shared rather than marketplace.
