+++ app/src/marketplace/i18n/index.js
@@ -1,21 +1,7 @@
 // @flow
 
-import merge from 'lodash/merge'
-import pick from 'lodash/pick'
-import sharedI18n from '$shared/i18n'
-import userpagesI18n from '$userpages/i18n'
-import authI18n from '$auth/i18n'
-
 import en from './en.po'
 
-const localI18n = {
+export default {
     en,
 }
-
-export default pick(merge(
-    {},
-    sharedI18n,
-    userpagesI18n,
-    authI18n,
-    localI18n,
-), Object.keys(localI18n))
